+++ src/Mobile.vue
@@ -4,22 +4,28 @@
       <PanelMain
         :background-color="config.backgroundColor"
         :broadcaster-name="broadcasterName"
+        :countdown-background-color="config.countdownBackgroundColor"
+        :countdown-font-color="config.countdownFontColor"
         :day-border-color="config.dayBorderColor"
         :font-color="config.fontColor"
         :font-family="config.fontFamily"
         :font-size="config.fontSize"
         :header-background-color="config.headerBackgroundColor"
         :header-font-color="config.headerFontColor"
+        :is-live="isLive"
         is-mobile
         :panel-title="config.panelTitle"
         :schedule-button-background-color="config.scheduleButtonBackgroundColor"
         :schedule-button-font-color="config.scheduleButtonFontColor"
         :schedule-items="schedule"
         :show-category="config.showCategory"
+        :show-countdown="config.showCountdown"
         :show-header="config.showHeader"
         :show-times="config.showTimes"
         :show-title="config.showTitle"
         :show-usernames="config.showUsernames"
+        :start-live-polling="startLivePolling"
+        :stop-live-polling="stopLivePolling"
         :time-font-color="config.timeFontColor"
         :vacation="vacation"
         :vacation-background-color="config.vacationBackgroundColor"
@@ -36,5 +42,5 @@ import PanelMain from './components/panel/PanelMain.vue';
 import CustomLoader from './components/common/CustomLoader.vue';
 import { useTwitch } from './composables/twitch.composable';
 
-const { broadcasterName, config, schedule, twitchLoading, vacation } = useTwitch();
+const { broadcasterName, config, isLive, schedule, twitchLoading, vacation, startLivePolling, stopLivePolling } = useTwitch();
 </script>